Add geolocation info in request

+++ db/migrate/20180426161502_create_ubs.rb
@@ -5,8 +5,8 @@ class CreateUbs < ActiveRecord::Migration[5.2]
       t.string :address
       t.string :city
       t.string :phone
-      t.decimal :geocode_lat
-      t.decimal :geocode_long
+      t.float :geocode_lat
+      t.float :geocode_long
       t.string :size
       t.string :adaptation_for_seniors
       t.string :medical_equipment
